requirements.txt
asgiref==3.5.2
attrs==22.1.0
certifi==2022.12.7
charset-normalizer==2.1.1
Django==4.1.4
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
drf-spectacular==0.24.2
idna==3.4
inflection==0.5.1
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.1
pdfkit==1.0.0
Pillow==9.3.0
psycopg2==2.9.5
pyrsistent==0.19.2
python-decouple==3.6
pytz==2022.6
PyYAML==6.0
requests==2.28.1
sqlparse==0.4.3
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.13
